+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 14-11-2023
 - configur for localhost to connect
 - add authentication
+- add basic profile details page 
 
 Once you are ready, visit "/users/register"
 to create your account and then access "/dev/mailbox" to